+++ limbo/mvvm/src/main/java/co/windly/limbo/mvvm/binding/ClickDataBinding.kt
@@ -15,7 +15,7 @@ interface ClickDataBinding {
     value = ["onThrottledClick", "throttleWindow"]
   )
   fun setOnThrottledClick(
-    view: View, onThrottledClick: OnClickListener, throttleWindow: Long = 1_000L)
+    view: View, onThrottledClick: OnClickListener, throttleWindow: Long)
 
   /**
    * Allows to register a long click listener to any view.